requirements.txt
numpy==1.26.4
opencv-python==4.10.0.84
deepface==0.0.93
pinecone-client==5.0.1
keras==3.6.0
tensorflow==2.17.0
tf_keras==2.17.0
matplotlib==3.9.2
